Create IndexwBTC.sol
I basically just followed along exactly what you had done previously, but the wBTC doesn't currently have a vault. IT just has the 'earn' location, so not sure if this would work for that. Or if having this in advance of a wbtc vault makes sense.

I also thought we could do a deposit into the crvBTC then deposit in the vault. That's a bit trickier, but I think doable.